Add files via upload
This project is based on Twitter API and medical concepts normalization. In this project, the data is collected from social media messages i.e., Twitter messages on medical concepts which enable several applications for enhancing the quality of people in a community. The data from social media can be leveraged to improve the understanding of patient's experiences in healthcare, such as the spread of infectious diseases and side effects of drugs. The frequent use of informal language, nonstandard grammar, and abbreviation forms, as well as social media messages, has to be taken into account by effective information extraction systems. To map a variable-length social media message to a medical concept in some external coding system is faced with a similar challenge in the task of medical concept normalization for social media text. In this work, we go beyond string matching. We propose to learn and exploit the semantic similarity between texts from social media messages and medical concepts using deep neural networks